requirements.txt
# Backend API
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0

# Base de datos
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# RabbitMQ
pika==1.3.2
# rabbitmq-client NO existe, es pika

# Cliente HTTP para APIs externas
requests==2.31.0

# Variables de entorno
python-dotenv==1.0.0

# Utilitarios
python-dateutil==2.8.2